Move 'user validation' code to a plugin directory
